move content widget in the widget folder
