--- ng2-components/ng2-activiti-form/src/components/activiti-content.component.css
@@ -1,46 +1,0 @@
-.upload-widget {
-    width: 100%;
-    word-break: break-all;
-}
-
-.upload-widget__content {
-    min-height: auto;
-}
-
-.upload-widget__icon {
-    float: left;
-    color: rgba(0, 0, 0, .26);
-}
-
-.upload-widget__file {
-    float: left;
-    margin-top: 4px;
-    color: rgba(0, 0, 0, .26);
-}
-
-.upload-widget__label {
-    color: rgba(0, 0, 0, .26);
-}
-
-.img-upload-widget {
-    width: 100%;
-    height: 100%;
-    border: 1px solid rgba(117, 117, 117, 0.57);
-    box-shadow: 1px 1px 2px #dddddd;
-    background-color: #ffffff;
-}
-
-.nothing-to-see {
-    box-shadow: 1px 1px 2px #dddddd;
-    background-color: #ffffff;
-    width: 100%;
-    height: 100%;
-    padding: 50px 0 50px 0;
-    text-align: center;
-}
-
-.previewTxt {
-    word-wrap: break-word;
-    word-break: break-all;
-    text-align: center;
-}
