+++ ng2-components/ng2-activiti-form/src/components/widgets/container/container.widget.scss
@@ -1,9 +1,6 @@
 @import 'theming';
 
 .adf {
-    .container-widget, .container-widget__header {
-    }
-
     &-field-list {
         padding: 0;
         list-style-type: none;
@@ -35,8 +32,5 @@ md-input-container {
     width: 95%;
 }
 
-md-grid-tile {
-    display: block;
-    position: absolute;
-    overflow: visible !important;
-}
+
+
