+++ ng2-components/ng2-activiti-form/src/components/widgets/unknown/unknown.widget.scss
@@ -1,0 +1,4 @@
+.adf-unknown-text{
+    margin-left: 10px;
+    color: red;
+}
